+++ packages/thumbprint-react/components/AlertBanner/index.tsx
@@ -18,7 +18,7 @@ interface AlertBannerPropTypes {
     dataTestId?: string;
     /**
      * A selector to hook into the React component for use in automated testing environments.
-     * @deprecated
+     * @deprecated Deprecated in favor of the `dataTestId` prop
      */
     dataTest?: string;
 }