+++ packages/thumbprint-react/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+-   [Minor] Add missing `dataTestId` prop to `Link` and `ThemedLink` components. (#724)
+
 ## 14.2.0 - 2020-10-21
 
 ### Added